connection.c: silent compiler warning

+++ src/microhttpd/connection.c
@@ -3377,7 +3377,7 @@ MHD_get_connection_info (struct MHD_Connection *connection,
     case MHD_CONNECTION_INFO_CONNECTION_SUSPENDED:
       return (const union MHD_ConnectionInfo *) &connection->suspended;
     case MHD_CONNECTION_INFO_CONNECTION_TIMEOUT:
-      connection->connection_timeout_dummy = connection->connection_timeout;
+      connection->connection_timeout_dummy = (unsigned int)connection->connection_timeout;
       return (const union MHD_ConnectionInfo *) &connection->connection_timeout_dummy;
     default:
       return NULL;
